feat(data.js): add static data for navigation, branding, categories, products, and footer links
This commit adds static data to the data.js file. This data includes navigation items, branding data, categories data, product data, and footer links. This data is necessary for the initial setup of the application and will be used to populate the various components of the application. This will ensure that the application has the necessary data to function correctly and provide a good user experience.